Filling blank 1040 with details
Added functionality to programmatically fill the blank 1040 form with the information that had been previously used to write to a new form. I was able to do this with the assistance of the 'pdftk' command line tool. For this to work, need to install the tool and place the .exe and .dll in either the /xampp folder, or your local test folder. Resources will be added to the bottom of the readme.